fix(map-maplibre): surface pre-extraction disk failures as a download state
startDownload() launched download() with no handler, so an IOException from
the disk work that runs before the extractor (clearing the previous region)
escaped the flow: downloadState stuck mid-way, and on Android an uncaught
exception in the repository's own SupervisorJob scope. Catch it there and
report Failed(IO_ERROR), matching how the extractor's and manifest's own
failures already surface. CodeRabbit follow-up on the same thread.

Also: ElevationTile compares size against width×height in Long, so a
product that wraps in Int (65536², say) can no longer match an empty array.
Ported identically to the Google-flavor branch.

+++ feature/map-terrain/src/commonMain/kotlin/org/meshtastic/feature/map/terrain/ElevationTile.kt
@@ -36,7 +36,8 @@ fun terrariumElevationMeters(red: Int, green: Int, blue: Int): Float =
 class ElevationTile(val width: Int, val height: Int, val elevations: FloatArray) {
     init {
         require(width > 0 && height > 0) { "width ($width) and height ($height) must both be positive" }
-        require(elevations.size == width * height) {
+        // Long math: width*height can wrap in Int (65536², say) and collide with a tiny elevations.size.
+        require(elevations.size.toLong() == width.toLong() * height) {
             "elevations.size (${elevations.size}) must equal width×height ($width×$height)"
         }
     }
